==> notebooks/2.0 - Datasets.ipynb <==
# Prep Work
To run this notebook properly on Colab you need to execute the following cell in order to clone the repository and get the necessary data

! git clone https://github.com/collab-uniba/Benchmark_sentimentAnalysis_ITA.git

# Datasets processing

## Import and preprocessing

import pandas as pd
pd.set_option('display.max_colwidth', None)
import warnings
warnings.filterwarnings("ignore")

#INPS
ht_inps=pd.read_csv('/content/Benchmark_sentimentAnalysis_ITA/data/raw/Enti/INPS/Hashtags.csv')
ht_inps['type'] = 'hashtag'
mn_inps=pd.read_csv('/content/Benchmark_sentimentAnalysis_ITA/data/raw/Enti/INPS/Mentions.csv')
mn_inps['type']='mention'

#INAIL
ht_inail=pd.read_csv('/content/Benchmark_sentimentAnalysis_ITA/data/raw/Enti/INAIL/Hashtags.csv')
ht_inail['type'] = 'hashtag'
mn_inail=pd.read_csv('/content/Benchmark_sentimentAnalysis_ITA/data/raw/Enti/INAIL/Mentions.csv')
mn_inail['type']='mention'

#Protezione Civile
ht_pc=pd.read_csv('/content/Benchmark_sentimentAnalysis_ITA/data/raw/Enti/Protezione Civile/Hashtags.csv')
ht_pc['type'] = 'hashtag'
mn_pc=pd.read_csv('/content/Benchmark_sentimentAnalysis_ITA/data/raw/Enti/Protezione Civile/Mentions.csv')
mn_pc['type']='mention'

import numpy as np

#INPS
#concatenation of hashtags and mentions dataframes indicating the type of tweet in a column
frames_inps = [ht_inps, mn_inps]
df_inps = pd.concat(frames_inps)
#column definition to identify retweets
df_inps['retweet'] = np.where(df_inps['tweet'].str.contains('RT @'), True, False)

#INAIL
#concatenation of hashtags and mentions dataframes indicating the type of tweet in a column
frames_inail = [ht_inail, mn_inail]
df_inail = pd.concat(frames_inail)
#column definition to identify retweets
df_inail['retweet'] = np.where(df_inail['tweet'].str.contains('RT @'), True, False)

#Protezione Civile
#concatenation of hashtags and mentions dataframes indicating the type of tweet in a column
frames_pc = [ht_pc, mn_pc]
df_pc = pd.concat(frames_pc)
#column definition to identify retweets
df_pc['retweet'] = np.where(df_pc['tweet'].str.contains('RT @'), True, False)

#Dataset infos
def get_stats(df):
    print("Dataset Shape: ",df.shape)
    print("\t Mentions - Hashtags")
    print("#Mentions:",df.loc[df['type'] == 'mention'].shape)
    print("#Hashtags:",df.loc[df['type'] == 'hashtag'].shape)
    print(df['type'].value_counts(normalize=True) * 100)
    if "retweet" in df:
        print("\t Retweet")
        print("#Retweet:",df.loc[df['retweet'] == True].shape)
        print(df['retweet'].value_counts(normalize=True) * 100)

get_stats(df_inps)

df_inps.head()

#Removing retweets and unnecessary columns

#INPS
df_inps=df_inps.loc[df_inps['retweet'] == False]
df_inps=df_inps[['created_at','tweet_id','tweet','type']]

#INAIL
df_inail=df_inail.loc[df_inail['retweet'] == False]
df_inail=df_inail[['created_at','tweet_id','tweet','type']]

#Protezione Civile
df_pc=df_pc.loc[df_pc['retweet'] == False]
df_pc=df_pc[['created_at','tweet_id','tweet','type']]

get_stats(df_inps)

# Silver labelling

#Emoji lists
positive_emoticons=["😀","😃","😄","😁","😆","🤣","😂","🙂","😊","😍","🥰","🤩","☺","🥳"]
negative_emoticons=["😒","😔","😟","🙁","☹","😥","😢","😭","😱","😞","😓","😩","😫","😡","😠","🤬"]

#Definition of silver labels based on the presence / absence of emojis within the entire tweet

#INPS
pos_df_inps = df_inps.loc[df_inps['tweet'].str.contains('|'.join(positive_emoticons))]
neg_df_inps = df_inps.loc[df_inps['tweet'].str.contains('|'.join(negative_emoticons))]
neutral_df_inps = pd.concat([df_inps, pos_df_inps, neg_df_inps]).drop_duplicates(keep=False)

#INAIL
pos_df_inail = df_inail.loc[df_inail['tweet'].str.contains('|'.join(positive_emoticons))]
neg_df_inail = df_inail.loc[df_inail['tweet'].str.contains('|'.join(negative_emoticons))]
neutral_df_inail = pd.concat([df_inail, pos_df_inail, neg_df_inail]).drop_duplicates(keep=False)

#Protezione Civile
pos_df_pc = df_pc.loc[df_pc['tweet'].str.contains('|'.join(positive_emoticons))]
neg_df_pc = df_pc.loc[df_pc['tweet'].str.contains('|'.join(negative_emoticons))]
neutral_df_pc = pd.concat([df_pc, pos_df_pc, neg_df_pc]).drop_duplicates(keep=False)

get_stats(neutral_df_pc)

#tweets containing both positive and negative emoticons
int_df_inps = pd.merge(pos_df_inps, neg_df_inps, how ='inner')
int_df_inail = pd.merge(pos_df_inail, neg_df_inail, how ='inner')
int_df_pc = pd.merge(pos_df_pc, neg_df_pc, how ='inner')

int_df_inps.shape

#Sampling neutral datasets to balance classes

#INPS
sample_neutral_df_inps = neutral_df_inps.sample(frac=0.015)

#INAIL
sample_neutral_df_inail = neutral_df_inail.sample(frac=0.015)

#Protezione Civile
sample_neutral_df_pc = neutral_df_pc.sample(frac=0.015)

neg_df_inail.shape

#Added polarity and topic column

#INPS
pos_df_inps['polarity']='positive'
pos_df_inps['topic']='inps'
neg_df_inps['polarity']='negative'
neg_df_inps['topic']='inps'
sample_neutral_df_inps['polarity']='neutral'
sample_neutral_df_inps['topic']='inps'

#INAIL
pos_df_inail['polarity']='positive'
pos_df_inail['topic']='inail'
neg_df_inail['polarity']='negative'
neg_df_inail['topic']='inail'
sample_neutral_df_inail['polarity']='neutral'
sample_neutral_df_inail['topic']='inail'

#Protezione civile
pos_df_pc['polarity']='positive'
pos_df_pc['topic']='pc'
neg_df_pc['polarity']='negative'
neg_df_pc['topic']='pc'
sample_neutral_df_pc['polarity']='neutral'
sample_neutral_df_pc['topic']='pc'

#concatenation of all dataframes
df_total = pd.concat([pos_df_inps, pos_df_inail, pos_df_pc,neg_df_inps,neg_df_inail,neg_df_pc,sample_neutral_df_inps,sample_neutral_df_inail,sample_neutral_df_pc])
df_total.head()

#Dataset shuffling
df_total_shuffle = df_total.sample(frac=1)
#Duplicate removal
df_total_shuffle = df_total_shuffle.drop_duplicates(keep='first',subset=['tweet_id'])

#First round annotation file
ann_1 = df_total_shuffle.sample(n=322,random_state=1)
ann_1.to_csv("/content/Benchmark_sentimentAnalysis_ITA/data/interim/Annotazione_1.csv")

#Second round annotation file
ann_2 = pd.concat([df_total_shuffle,ann_1]).drop_duplicates(keep=False,subset=['tweet_id'])
ann_2.to_csv("/content/Benchmark_sentimentAnalysis_ITA/data/interim/Annotazione_2.csv")
