==> notebooks/5.0 - Lexicon.ipynb <==
# Prep Work
To run this notebook properly on Colab you need to execute the following cell in order to clone the repository and get the necessary data

! git clone https://github.com/collab-uniba/Benchmark_sentimentAnalysis_ITA.git

# From Sentix to Sentistrenght

## Sentix lexicon import and processing

import pandas as pd

sentix = pd.read_csv('/content/Benchmark_sentimentAnalysis_ITA/data/external/lexicon/sentix',sep='\t',names=["Lemma","POS","Wordnet SynsetID","Pos Score","Neg Score","Polarity","Intensity"])
sentix.head()

sentix = sentix.dropna()
#Actual size of the lexicon
sentix.shape

#New polarity computation
sentix['New_Pol'] = sentix['Polarity']*sentix['Intensity']

sentix.head()

## Polysemy management

#Terms like 'buono' have more than one record in the lexicon
sentix[sentix['Lemma']=='buono']

#Get values for POS tagging used in the lexicon
uniqueValues = (sentix['POS']).unique()
uniqueValues

#Change POS values used in the lexicon to POS values used in Spacy for further computation
sentix['POS'] = sentix['POS'].replace(['a'],'adj')
sentix['POS'] = sentix['POS'].replace(['n'],'noun')
sentix['POS'] = sentix['POS'].replace(['r'],'adv')
sentix['POS'] = sentix['POS'].replace(['v'],'verb')

#Setting New_Lemma
sentix['New_Lemma'] = sentix['Lemma'] + '_' + sentix['POS']

sentix.head()

#Keep first entry for each Lemma_POS because, in WordNet, it has the highest absolute frequency
sentix = sentix.drop_duplicates(subset=['New_Lemma'],keep='first')

# Mapping Sentix New Polarity to SentiStrenght Polarity score

from sklearn.preprocessing import MinMaxScaler
pd.options.mode.chained_assignment = None  

#direct mapping of pos score, neg score and polarity through scaling
def sentix_to_sentistrenght (df):
    #New df deprived of useless columns for the purpose of converting to SentiStrenght format
    sentix_ss = df[['Lemma','New_Lemma', 'New_Pol','Pos Score','Neg Score','Polarity','Intensity']]
    #Separate dataframes for proper scaling
    sentix_ss_pos = sentix_ss[sentix_ss['New_Pol']>0]
    sentix_ss_neg = sentix_ss[sentix_ss['New_Pol']<0]
    sentix_ss_neut = sentix_ss[sentix_ss['New_Pol']==0]

    #Positive dataframe
    #Min max scaling on polarity scores
    scaler = MinMaxScaler(feature_range=(2,5))
    sentix_ss_pos[['Polarity SS']] = scaler.fit_transform(sentix_ss_pos[['New_Pol']])
    #Rounding of scaled scores
    sentix_ss_pos[['Polarity SS']] = sentix_ss_pos[['Polarity SS']].round()

    #Negative dataframe
    #Polarità negativa
    sentix_ss_neg['New_Pol'] = sentix_ss_neg['New_Pol'].abs()
    #Min max scaling on polarity scores
    scaler = MinMaxScaler(feature_range=(2,5))
    sentix_ss_neg[['Polarity SS']] = scaler.fit_transform(sentix_ss_neg[['New_Pol']])
    #Rounding of scaled scores
    sentix_ss_neg[['Polarity SS']] = sentix_ss_neg[['Polarity SS']].round()
    #Return the values to their original negativity
    sentix_ss_neg['Polarity SS'] = -sentix_ss_neg['Polarity SS'].abs()
    sentix_ss_neg['New_Pol'] = -sentix_ss_neg['New_Pol'].abs()

    #Neutral dataframe
    sentix_ss_neut['Polarity SS'] = (sentix_ss_neut['New_Pol'])

    #Final dataframe mapped from Sentix to Sentistrenght
    sentix_ss_total = pd.concat([sentix_ss_pos, sentix_ss_neg, sentix_ss_neut], axis=0)
    sentix_ss_total['Polarity SS'] = sentix_ss_total['Polarity SS'].round()
    sentix_ss_total = sentix_ss_total[['Lemma','New_Lemma','New_Pol','Polarity SS','Pos Score','Neg Score','Polarity','Intensity']]

    return sentix_ss_total

sentix_SS = sentix_to_sentistrenght(sentix)

sentix_SS[sentix_SS['Lemma']=='buono']

sentix_SS.to_csv("/content/Benchmark_sentimentAnalysis_ITA/data/external/lexicon/sentix_ss.csv")
