==> notebooks/7.0 - Bert.ipynb <==
# Prep Work
To run this notebook properly on Colab you need to execute the following cell in order to clone the repository and get the necessary data

! git clone https://github.com/collab-uniba/Benchmark_sentimentAnalysis_ITA.git

# Environment setup

#Requirements
! pip install spacy==3.1.1
! pip install spacy-transformers
! python -m spacy download it_core_news_lg
! pip install typing_extensions

import pandas as pd
from datetime import datetime
import spacy
import spacy_transformers
from spacy.tokens import DocBin

# Dataset

## Enti

# Reading the dataset
df = pd.read_csv("/content/Benchmark_sentimentAnalysis_ITA/data/processed/Gold_Standard.csv")
#Mapping GoldStandard scores to sentiment classes
df.loc[df['Class'] =='pos', 'Pos_GS'] = 'yes'
df.loc[df['Class'] == 'neg', 'Neg_GS'] = 'yes'
df.loc[df['Class'] == 'neut', 'Neut_GS'] = 'yes'
df.loc[df['Class'] == 'mix', 'Pos_GS'] = 'yes'
df.loc[df['Class'] == 'mix', 'Neg_GS'] = 'yes'
df = df.fillna("no")

df.head()

df['tweet_x'][0]

df.shape

#Train-test splitting
train = df.sample(frac = 0.8, random_state = 25)
test = df.drop(train.index)

# Checking the shape
print(train.shape, test.shape)

## SentiPolc

sentiTrain =  pd.read_csv("/content/Benchmark_sentimentAnalysis_ITA/data/processed/SentiPolc/training_set_sentipolc16.csv")
sentiTrain.loc[sentiTrain['opos'] ==1, 'Pos_GS'] = 'yes'
sentiTrain.loc[sentiTrain['oneg'] ==1, 'Neg_GS'] = 'yes'
sentiTrain.loc[(sentiTrain['lpos'] ==0) & (sentiTrain['lneg'] ==0), 'Neut_GS'] = 'yes'
sentiTrain = sentiTrain.fillna("no")

colnames=['subj','opos','oneg','iro','lpos','lneg','top','text'] 
sentiTest = pd.read_csv("/content/Benchmark_sentimentAnalysis_ITA/data/processed/SentiPolc/test_set_sentipolc16_gold2000.csv",header=None,names=colnames)
sentiTest.loc[sentiTest['opos'] ==1, 'Pos_GS'] = 'yes'
sentiTest.loc[sentiTest['oneg'] ==1, 'Neg_GS'] = 'yes'
sentiTest.loc[(sentiTest['opos'] ==0) & (sentiTest['oneg'] ==0), 'Neut_GS'] = 'yes'
sentiTest = sentiTest.fillna("no")

import spacy
nlp = spacy.load("it_core_news_lg")
nlp.pipe_names

## Spacy input format conversion

def document(data):
  text = []
  for doc, label in nlp.pipe(data, as_tuples = True):
    if (label=='yes'):
      doc.cats['yes'] = 1
      doc.cats['no'] = 0
    elif (label=='no'):
      doc.cats['yes'] = 0
      doc.cats['no'] = 1
    text.append(doc)
  return(text)

def convertToSpacyEnti (sentiment,train,test):
    if sentiment=='pos':
        train['tuples'] = train.apply(lambda row: (row['tweet_x'],row['Pos_GS']), axis=1)
        train = train['tuples'].tolist()
        test['tuples'] = test.apply(lambda row: (row['tweet_x'],row['Pos_GS']), axis=1)
        test = test['tuples'].tolist()
    elif sentiment=='neg':
        train['tuples'] = train.apply(lambda row: (row['tweet_x'],row['Neg_GS']), axis=1)
        train = train['tuples'].tolist()
        test['tuples'] = test.apply(lambda row: (row['tweet_x'],row['Neg_GS']), axis=1)
        test = test['tuples'].tolist()
    elif sentiment=='neut':
        train['tuples'] = train.apply(lambda row: (row['tweet_x'],row['Neut_GS']), axis=1)
        train = train['tuples'].tolist()
        test['tuples'] = test.apply(lambda row: (row['tweet_x'],row['Neut_GS']), axis=1)
        test = test['tuples'].tolist()
    #Time for converting into binary document for train dataset
    start_time = datetime.now()
    #Passing the train dataset into function 'document'
    train_docs = document(train)
    #Creating binary document using DocBin function in spaCy
    doc_bin = DocBin(docs = train_docs)
    #Saving the binary document as train.spacy
    doc_bin.to_disk("/content/Benchmark_sentimentAnalysis_ITA/data/processed/Enti/train_"+sentiment+".spacy")
    #Time duration for train dataset
    end_time = datetime.now()
    print('Training dataset creation: {}'.format(end_time - start_time))
    #Time for converting into binary document for test dataset
    start_time = datetime.now()
    #passing the test dataset into function 'document'
    test_docs = document(test)
    doc_bin = DocBin(docs = test_docs)
    doc_bin.to_disk("/content/Benchmark_sentimentAnalysis_ITA/data/processed/Enti/valid_"+sentiment+".spacy")
    #Printing the time duration for test dataset
    end_time = datetime.now()
    print('Test dataset creation: {}'.format(end_time - start_time))
    

def convertToSpacySentiPolc (sentiment,train,test):
    if sentiment=='pos':
        train['tuples'] = train.apply(lambda row: (row['text'],row['Pos_GS']), axis=1)
        train = train['tuples'].tolist()
        test['tuples'] = test.apply(lambda row: (row['text'],row['Pos_GS']), axis=1)
        test = test['tuples'].tolist()
    elif sentiment=='neg':
        train['tuples'] = train.apply(lambda row: (row['text'],row['Neg_GS']), axis=1)
        train = train['tuples'].tolist()
        test['tuples'] = test.apply(lambda row: (row['text'],row['Neg_GS']), axis=1)
        test = test['tuples'].tolist()
    elif sentiment=='neut':
        train['tuples'] = train.apply(lambda row: (row['text'],row['Neut_GS']), axis=1)
        train = train['tuples'].tolist()
        test['tuples'] = test.apply(lambda row: (row['text'],row['Neut_GS']), axis=1)
        test = test['tuples'].tolist()
    #Time for converting into binary document for train dataset
    start_time = datetime.now()
    #Passing the train dataset into function 'document'
    train_docs = document(train)
    #Creating binary document using DocBin function in spaCy
    doc_bin = DocBin(docs = train_docs)
    #Saving the binary document as train.spacy
    doc_bin.to_disk("/content/Benchmark_sentimentAnalysis_ITA/data/processed/SentiPolc/train_"+sentiment+".spacy")
    #Time duration for train dataset
    end_time = datetime.now()
    print('Training dataset creation: {}'.format(end_time - start_time))
    #Time for converting into binary document for test dataset
    start_time = datetime.now()
    #passing the test dataset into function 'document'
    test_docs = document(test)
    doc_bin = DocBin(docs = test_docs)
    doc_bin.to_disk("/content/Benchmark_sentimentAnalysis_ITA/data/processed/SentiPolc/valid_"+sentiment+".spacy")
    #Printing the time duration for test dataset
    end_time = datetime.now()
    print('Test dataset creation: {}'.format(end_time - start_time))

convertToSpacyEnti('pos',train,test)
convertToSpacyEnti('neg',train,test)
convertToSpacyEnti('neut',train,test)

convertToSpacySentiPolc('pos',sentiTrain,sentiTest)
convertToSpacySentiPolc('neg',sentiTrain,sentiTest)
convertToSpacySentiPolc('neut',sentiTrain,sentiTest)

# SpaCy
The next cells allows you to create the configuration file and train the model using SpaCy. To modify the hyperparameters and the train and test data it is necessary to access the configuration file config.cfg (first parameter passed to the spacy train function). For more complete information on this topic you can refer to this [page](https://spacy.io/usage/training). To change the directory in which to save the results, it is necessary to modify the output parameter of the spacy train function, i.e. the path immediately following the wording --output
## SpaCy config file

#Converting base configuration into full config file

! python -m spacy init fill-config /content/Benchmark_sentimentAnalysis_ITA/models/configs/base_config.cfg /content/Benchmark_sentimentAnalysis_ITA/models/configs/standard_config.cfg

## SpaCy training

start_time = datetime.now()
 
! python -m spacy train "/content/Benchmark_sentimentAnalysis_ITA/models/configs/config.cfg" --verbose  --output "/content/Benchmark_sentimentAnalysis_ITA/models/Models/BERT-xxl/Within_E/Negative" --gpu-id 0

end_time = datetime.now()

print('Duration: {}'.format(end_time - start_time))
